Add files via upload
I have created a game using GUI(tkinter module and pygame for sound effects) where there will be a triangle of some coordinates and there will  be a ball the user needed to shoot. So the user needed to enter the coordinate then the ball will shoot out to that coordinate and if the ball shoots out inside the triangle then the user will increment 100 as score or else the score will be deducted with -50. So the user need to ensure the ball shoots inside so that the user will get good score. I have created three levels in the game easy ,medium, Hard, Even here the user will have a timer for each level.